refactor: 디렉토리 생성 및 구조 변경
