+++ src/main/java/lotto/Main.java
@@ -1,5 +1,7 @@
 package lotto;
 
+import lotto.controller.LottoMachine;
+
 public class Main {
     public static void main(String[] args) {
         LottoMachine lottoMachine = new LottoMachine();
