+++ src/main/java/lotto/model/WinningLotto.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.model;
 
 public class WinningLotto {
     final int matchedCount;
