+++ src/main/java/lotto/view/Output.java
@@ -1,6 +1,6 @@
-package lotto.ui;
+package lotto.view;
 
-import lotto.Lotto;
+import lotto.model.Lotto;
 
 import java.util.List;
