+++ src/test/java/lotto/LottoTest.java
@@ -1,9 +1,10 @@
 package lotto;
 
+import lotto.controller.LottoMachine;
+import lotto.model.Lotto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
 
 import java.util.ArrayList;
